+++ generators/mwc16_32.c
@@ -19,7 +19,9 @@ static inline uint64_t get_bits_raw(void *state)
 {
     Lcg32State *obj = state;
     uint16_t x_lo = obj->x & 0xFFFF, x_hi = obj->x >> 16;
-    obj->x = 63885 * x_lo + x_hi;
+    obj->x = (uint32_t)63885 * x_lo + x_hi;
+//  Causes a glitch in GCC 10 (some UB during optimization?)
+//    obj->x = 63885 * x_lo + x_hi; 
     return obj->x;
 }
